Add cli alias for easier program accessibility
This change allows the user to call simpleSize from any directory
within the file system. To use, call

simplesize [target directory path] [-u UNIT]

The user can use relative and absolute paths to specify the target
directory. If a relative path was provided, the path is relative to
the directory from which the user is calling this script.